Tighten frontend/emitter semantics before CLI regressions

+++ packages/emitter/testcases/common/expected/edge-cases/inline-object-param/InlineObjectParam.cs
@@ -8,6 +8,6 @@ namespace TestCases.common.edgecases.inlineobjectparam
             return x.a;
         }
 
-        public static readonly int result = takes(new global::TestCases.common.edgecases.inlineobjectparam.__Anon_d6a3_aaa895b9 { a = (int)1, b = "hi" });
+        public static readonly int result = takes(new global::TestCases.common.edgecases.inlineobjectparam.__Anon_d6a3_aaa895b9 { a = 1, b = "hi" });
     }
 }
