+++ packages/emitter/testcases/common/expected/types/expected-type-threading/VariableInit.cs
@@ -3,10 +3,10 @@ namespace TestCases.common.types.expectedtypethreading
     [global::Tsonic.Internal.ModuleContainerAttribute]
     public static class VariableInit
     {
-        public static readonly int value = (int)100;
+        public static readonly int value = 100;
 
-        public static readonly int a = (int)10;
+        public static readonly int a = 10;
 
-        public static readonly int b = (int)20;
+        public static readonly int b = 20;
     }
 }
